docker: Allow parallel pulls of different images
Before this patch, we shoved a giant lock around the entire pull
function.  This effectively meant that we had to pull images serially
taking considerably longer that we'd like.  This patch parallelizes
the code so that different images can be downloaded concurrently,
while attempts to download the same image remain serial.