ofctl: Dramatically simplify the OpenFlow code
This patch relies on several techniques to dramatically simplify the
OpenFlow code generated in worker.go.  First, it relies on the
`ovs-ofctl replace-flows` to do the join, instead of the minion.  This
relieves the minion from needing to parse flows carefully to ensure a
successful join.  Second, it takes advantage of multiple OpenFlow
tables to reduce the number of per-container flows necessary.  And
finally, third, it relies on the NAT to filter public internet traffic
instead of doing it in OpenFlow.