cluster: Implement a two stage machine join
The machine join can be rather fuzzy -- multiple database machines may
match a particular cloud machine to varying degrees of fit.  Before
this patch, in some cases it was possible for the join to make poor
(but barely legal) match that preempted a perfect match that could
have been made later.  This patch resolves the issue by doing
the machine join in two stages.  First, it attempts to pair those
machines that match perfectly, before falling back to a fuzzier match
for the leftovers.