minion: Join containers based exclusively on the StitchID
Now that the StitchID is more carefully crafted by Quilt.js, it seems
logical (and makes the system easier to reason about) if a given
container always maintains the same StitchID even as the spec changes.

This fixes a bug where booting a second copy of a particular system on
the same infrastructure could cause new nodes to connect to the old
system.