haproxy: Reimplement dns resolution at startup
The way we currently configure HAProxy, it resolves the hostnames of
its backend servers once at startup.  If that DNS resolution times out
because OVS is not yet configured, HAProxy will crash.  The correct
long term solution to this problem is to configure HAProxy to resolve
host names at runtime as necessary.  However, until that's
implemented, this patch resurrects old DNS resolution code that was
pulled out originally because it didn't play nicely with /etc/hosts,
but should now work.