ovsdb: Lightly refactor OVSDB while transitioning to mockery
The fake ovsdb client library implemented in previous patches was
massive and extremely difficult to understand. This patch migrates the
tests for the ovsdb library, and its callers, to mockery instead.  In
order to aid unit testing, it does some light refactoring of the ovsdb
library as well.  Note that this patch neglects to cover some
previously tested functions that are deleted in future patches.