plugin: Create veth pair in CreateEndpoint()
This patch moves veth pair creation from Join() to CreateEndpoint().
The network plugin design seems to imply it should be created there,
and looking through some of the official docker plugins, that's indeed
where they do it.  Furthermore, future patches require the veth to be
created in CreateEndpoint() for OpenFlow rules to be installed in the
plugin.