amazon: Stop tagging spot requests
When the amazon provider was first written, there was no way to filter
spot requests based on the security group they were in.  So, instead,
it tagged spot requests with their namespace in lieu of checking the
security group directly.

Not only did this approach complicate the code, it also lead to a
brief window of time where a spot request was created, but not yet
tagged.  During this window all quilt daemons running on a particular
aws account would consider those spot requests as superfluous members
of in their own namespace, and would thus cancel them.  This patch
does away with this issue, and simplifies the code, by simply
filtering spot requests on their security group when listed.