cloud: Re-architect the cloud package
Before this patch, the database machine table contained the set of
desired machines as populated by the engine.  And the cloud package
was responsible for implementing that specification. This had a number
of problems including:

    * Blueprint changes were reflected in the database and `quilt ps`
    immediately, before they had time to be enacted.  This confused
    users.

    * It was unnecessarily complex -- especially the `engine` package
    that had a confusing relationship with the cloud package.

    * The cloud joins were unnecessarily complicated as they had to
    carefully blend the desired machines with information from what
    was actually running.

This patch re-architects the cloud package with a new approach.  It
now populates the database machine table with the set of machines that
are running in the cloud provider.  And it joins that machine table
with the set of desired machines as retrieved from the database
blueprint table.  As a result:

    * Machines have a new "Stopping" status that shows in `quilt ps`
    when machines are being terminated.

    * The machine BlueprintID is no longer used, in favor of the ID
    provided by the cloud provider.  This even applies to the cli
    where users can cross-reference the IDs we show them with what
    their cloud provider knows about.

    * Numerous other code deletions, simplifications, and cleanups.