tests: Refactor the network test
Before this patch, the network test code was extremely complicated and
slow.  SSH attempts tended to come in batches, causing the system to
either bottleneck on the concurrency limit, or waste time trying to
figure out what to do next.  This patch refactors the test and in the
process dramatically simplifies the code. It additionally causes SSH
executions to smooth out at a rate of once every 100ms which results
in a roughly 3x improvement in test time in my local environment.
This makes the network test much more palatable as a development tool.