network: Add support for address sets
As spark scaled, the number of ACLs that OVN had to process grew at
O(n^2) as the number of workers grew.  At a scale of 64 nodes this
would cause OVN to spend several minutes processing ACLs before the
network would work properly.

This patch dramatically reduces the number of ACLs by adding support
for address sets to the Kelda Daemon.  Address sets allow ACLs to be
express groups of IP addresses, instead of requiring them to be
exploded out into a separate ACL per IP.

This patch adds full support for address sets to the deployment
engine, and partial support to the Javascript bindings.  Even with
partial support for address sets, ACLs are no longer a bottleneck when
running Spark.  However, full support will need to be added later if
we want to scale to large numbers of containers.