google: Refactor the network architecture
This patch rewrites most of the Google network related code in service
of several goals.  First, it creates a Google network per region
rather than requiring all regions to share a single network.  This
facilitates us, second, cleaning up the Google network in regions that
are unused.  Third, it generally simplifies the ACL code such that
it's more readable, less redundant, and easier to test.  And finally,
it adds unit tests for the sections of the code that it updates.