google: Add filtering for all List functions
This patch adds a mechanism to filter all List functions in the google
client so they only return results that will be relevant to that
particular provider.  It also changes ListInstances() to take a
description instead of a filter, making it easier to use and more
consistent with the other List functions modified in this patch.