kelda: Replace scheduler with Kubernetes
Before, we had a custom scheduler for ensuring the desired containers
were running. While it is not too difficult to implement a basic
scheduler, moving to Kubernetes will allow us to iterate more quickly on
features such as volumes since Kubernetes already implements much of the
logic for actually creating and managing volumes -- we will be able to
just focus on how to express volumes in the blueprint language.

Furthermore, this change allows us to potentially, in the future, run on
top of an existing Kubernetes cluster. This will be helpful given the
general movement in industry towards using Kubernetes.

At a high level, this commit encompasses four major changes:
1) It modifies the deployment engine to boot a Kubernetes cluster on the
Kelda machines.
2) It replaces the custom scheduler with code that translates the
containers specified by the user into deployments to be enforced by the
Kubernetes cluster.
3) It replaces the Docker network plugin with a CNI network plugin
compatible with Kubernetes.
4) It replaces Vault in favor of the built-in Kubernetes solution for
secret management. It's simpler to just use the Kubernetes solution, and
we weren't making use of any of the more advanced Vault features.