scheduler: Write placements in terms of hostnames
We're phasing out blueprint IDs. This also fixes a bug in the Kubernetes
scheduler where a container's blueprint ID changes (because an
attribute such as filepathToContent changes), but the hostname doesn't
change. Before, when we tagged pods with their blueprint ID via a label
this would require updating the pod's labels, which isn't supported by
Kubernetes.