docker: Pull hyperkube into the minion docker image
In future patches, the kubelet container will need to access the ovs
utilities to properly configure the network.  Instead of adding the
ovs binaries to the kubelet image, this patch pulls the hyperkube
binary into the main minion docker image, and runs kubelet from that.

As part of this process, this path puts the docker image path into a
global constant, so that it doesn't have to be hard-coded in multiple
places.