cloud: Prefer to leave a floating IP in place when possible
Before this patch, the cloud preferred to join machines with known
roles even if that meant reassigning a floating IP away from a machine
with an unknown role to a different worker with a known role.  This
meant that on boot (or namespace change), a perfectly health machine
could have its floating IP taken away from it, simply because one of
its peers declared its role to the daemon before it got a chance to.

There's really no particular reason for this preference, so this patch
fixes the issue by reversing it.  All things being equal, we will
prefer to match a machine with the correct floating IP, even if that
means its role is unknown.